Update README.md
Business Problem: Our task is to analyze the dataset and predict whether the income of an adult will exceed 50k per year or not


CONCLUSION: In first part i didn't get the good result(accuracy) due to scaling of data. So,in second part i build a model with unscaled dataset, it gives good accuracy.